requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
streamlit-ace>=0.1.1
langchain>=0.1.0
langchain-community>=0.0.10
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
PyPDF2>=3.0.1
python-docx>=0.8.11
markdown>=3.5.1 